fix: break circular import from get_max_layer_num consolidation
The consolidation made the converting_*.py modules import trainer/weights.py,
but weights.py imports trainer/lora.py, which (via trainer/models/__init__)
imports the model modules and their converting_*.py — closing a cycle:

  lora -> models/__init__ -> afmoe/modeling -> converting_afmoe
    -> weights -> lora (partially initialized) -> ImportError

Move get_max_layer_num into a dependency-free leaf module
trainer/conversion_utils.py (imports only torch) and point the seven
converting modules + the NCCL broadcaster at it. weights.py no longer
defines it (it never used it internally). Keeps the single-source dedup
without the cycle.

Caught by the reverse_text integration tests (trainer crashed at import).
